First Commit. A clean build, centered around harmony_state
harmony_state is now the central tool. The generative needle on the phonograph.

The other tools can be part of packages and graphing, but should not be necessary -- not much.

The ability to ouput in the terminal as a stream remains important. It was once done but has since been lost.

Needs to be tested as a package.